feat: added encrypted username field to blocks collection

+++ apps/web/src/hooks/useBlocks.ts
@@ -1,45 +1,43 @@
 import { useEffect, useState } from "react";
 import { useDB } from "@/hooks/useDB";
 import { BlockType } from "@chat/core";
+import { useAuth } from "./useAuth";
 
 export function useBlocks() {
-    const { db } = useDB();
-    const [blocks, setBlocks] = useState<BlockType[]>([]);
+  const { db, getAllDecr } = useDB();
+  const { key } = useAuth();
+  const [blocks, setBlocks] = useState<BlockType[]>([]);
 
-    useEffect(() => {
-        if (!db) return;
+  useEffect(() => {
+    if (!db || !key) return;
 
-        const setup = async () => {
-            const allblocks = await db.getAll("blocks");
-            console.log(allblocks)
-            const mapped = allblocks.map(block => ({ userId: block.userId } as BlockType));
-            setBlocks(mapped);
-        };
+    (async () => {
+      const blocks = (await getAllDecr("blocks", key)) as BlockType[];
+      setBlocks(blocks);
+    })();
+  }, [db, key]);
 
-        setup();
-    }, [db]);
+  const unblock = async (block: BlockType) => {
+    if (!db || !key) return;
 
-    const unblock = async (block: BlockType) => {
-        if (!db) return;
+    const tx = db.transaction("blocks", "readwrite");
+    const store = tx.objectStore("blocks");
 
-        const tx = db.transaction("blocks", "readwrite");
-        const store = tx.objectStore("blocks");
+    let cursor = await store.openCursor();
 
-        let cursor = await store.openCursor();
+    while(cursor) {
+      const _block = cursor.value;
+      if (_block.userId === block.userId) {
+        await cursor.delete();
+      }
+      cursor = await cursor.continue();
+    }
 
-        while(cursor) {
-            const _block = cursor.value;
-            if (_block.userId === block.userId) {
-                await cursor.delete();
-            }
-            cursor = await cursor.continue();
-        }
+    await tx.done;
+    setBlocks((prev) => prev.filter((b) => b.userId !== block.userId));
+    const allblocks = (await getAllDecr("blocks", key)) as BlockType[];
+    console.log(allblocks)
+  };
 
-        await tx.done;
-        setBlocks((prev) => prev.filter((b) => b.userId !== block.userId));
-        const allblocks = await db.getAll("blocks");
-        console.log(allblocks)
-    };
-
-    return { blocks, unblock };
+  return { blocks, unblock };
 }
